setbuff instead of remove, added hook to deflect
